// src/app/modules/matenimiento/models/salidaMaterial/responseListVWSalidaMaterial.model.ts
export class ResponseListSalidaMaterial {
    cantidad: number= 0;
    idMaterial: number= 0;
    idProduccion: number= 0;
    codigoProduccion :string="";
    nombreMaterial: string="";
    idUnidad: number=0;
    fechaSalida: string="";
    idSalidaMaterial: number=0;
    estado: string ="";
    stock: number= 0;
}